Added missing files, and fixed some bugs - Added files that were supposed to be added a few commits back (doh!) - Fixed setting position of a sprite - Fixed PNG loading - Font sizes can be fractions now

+++ libs/Rainbow/FontAtlas.h
@@ -28,9 +28,9 @@
 class FontAtlas
 {
 public:
-	const unsigned int pt;
+	const float pt;
 
-	FontAtlas(const char *font_family, const unsigned int pt);
+	FontAtlas(const char *font_family, const float pt);
 	~FontAtlas() { }
 
 	/// Prints text at (x,y).
